refactor: remove excess table
